fix: Split searchbox into several compound components
